+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## v0.0.3 - to be released
+## v0.0.4 - 2022-10-17
+
+### Fixed
+
+- Toisto did not start because the rich library was not bundled.
+
+## v0.0.3 - 2022-10-17
 
 ### Added